Refactors auctions route
This pushes a lot of code up onto the mode. Creating an addBid method where much of the validation occurs when a user attempts to add a bid.

In this case any errors encountered are thrown as errors which are returned to the calling route.

The route is then able to use try catch to return the error as an API response when the async/await call fails.