Adds the addBid method to Auction model
This runs all necessary validations by calling the related methods on the object instance.

If the bid fails at any point we throw a validation error.

This also checks for existing Autobids in the system. If one of these exists then that is used as the incremented bid, with an error being returned to the current bidder saying they were already outbid.