Fixes possible exploit
Uses the User ID decoded from the token sent in the request.

Previously this code was using the User ID sent in the request body, this could lead to exploitation, as long as the logged in user had a valid token they could send another person's User ID in the request instead.